// src/app/enum/tipo-exercicio-imagens.enum.ts
export enum ETipoExercicioImagens{
	PALAVRA_ORGANIZADAS = 'Arrastar palavra organizadas até sua imagem',
	PALAVRA_DESORGANIZADAS = 'Arrastar palavra desorganizadas até sua imagem',
	PALAVRA_DESORGANIZADAS_2 = 'Arrastar palavra desorganizadas até sua imagem (2 imagens)',
	PALAVRA_ORGANIZADAS_2 = 'Arrastar palavra organizadas até sua imagem (2 imagens)',
}