// src/app/enum/tipo-exercicio-silaba.enum.ts
export enum ETipoExercicioSilaba {
	SILABAS_ORGANIZADAS_ESCRITA = "Arrastar sílabas organizadas até sua sombra(Escrita)",
	SILABAS_DESORGANIZADAS_ESCRITA = 'Arrastar sílabas desorganizadas até sua sombra(Escrita)',
	SILABAS_ORGANIZADAS_SOM = 'Arrastar sílabas organizadasaté sua sombra(Som)', // exercicio tipo 5
	SILABAS_DESORGANIZADAS_SOM = 'Arrastar sílabas desorganizadasaté sua sombra(Som)', // exercicio tipo 5
	SILABAS_EMBARALHADAS = 'Silabas ficam embaralhadas(Embaralhado)',
}